requirements.txt

alpaca-trade-api==3.0.0
aiohttp==3.8.4
arxiv==1.4.3
beautifulsoup4==4.11.2
chromadb==0.4.3
google-api-python-client==2.80.0
kafka-python==2.0.2
numpy==1.24.2
openai==0.27.8
pandas==1.5.3
pyspark==3.2.0
python-dotenv==1.0.0
redis==4.5.1
scholarly==1.7.11
websockets==10.4

